requirements.txt
 django-ckeditor